pyproject.toml
[tool.poetry]
name = "electronics-retail-chain"
version = "0.1.0"
description = ""
authors = ["Averyanova Elizaveta"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
